Add files via upload
Run the XAMPP control panel and start MySQL and Apache
Go to C:\xampp\htdocs and extract the downloaded zip file (healthcare) inside the folder
Open the browser and go to http://localhost/phpmyadmin/ to create the database
Click the new to create a database.
Name the database sourcecodester_healthcaredb.
Click Import to import the SQL file. Click choose file and select the file that can be found inside the healthcare folder
Click go.